Changed global logic of meassurements
Removed the daily averages. It provided not really
good information and was not really ben used. In
addition this added a lot of complexity to the tool.

Then, the taken date was confused. The tool was
expecting this is the data when something happened.
And therefore, it had to go back and forward depending
on taking daily or hourly. To complex.

The new logic is independent on the dates when something happenns.
It just takes one day, and do hourly averages from there
up to the next X configured days. What happened during
these days is completely out of the tool.

The person that executes the tool, knows what happened
during some of the days. And can use the day-labels
to show in the graphics what happened there.